more small changes
